backend/requirements.txt
flask>=2.0.1
openai>=1.0.0
python-docx>=0.8.11
pypdf2>=3.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0